requirements.txt
streamlit
python-dotenv
langchain-groq
langchain-community
